requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
